Add support for a non-square canvas.
The original code assumed a square canvas and would distort when trying to
draw to a rectangle.

To fix this, the idea of a pixel_size is introduced to have a square pixel and
allow a rectangle viewport.  This breaks the 1-1 square assumption as now the
x and y axes don't have to try and scale to the same value.